+++ CHANGELOG.md
@@ -190,6 +190,7 @@
 165. Fixed Bad Image Format Exception Issue with Loading DLLs from ZIPs.
 166. Fixed Issue with Assembly Generator Contacting All RemoteAPI Hosts when One Responded.
 167. Added OnSceneWasUnloaded MelonMod Override.
+168. Fixed IOException Issue with Assembly Generator extracting ZIP Archives.  (Credits to Extacy :D)
 
 ---